remove flask, marshmallow & gunicorn
