+++ requirements.txt
@@ -2,9 +2,7 @@ uvloop==0.14.0
 PyYAML==5.1.2
 rdflib==4.2.2
 rdflib-jsonld==0.4.0
-gunicorn==19.7.1
 protobuf==3.11.0
 click==7.0
 fastapi==0.44.1
 uvicorn==0.10.8
-requests==2.22.0
