+++ setup.py
@@ -1,42 +1,8 @@
 # setup.py
 # Author: Thomas MINIER - MIT License 2017-2018
 from setuptools import setup, find_packages
-from subprocess import run, SubprocessError
-from setuptools.command.develop import develop
-from setuptools.command.install import install
 from os import getcwd
 from sys import exit
-
-
-def install_web_deps():
-    """Install Sage web interface dependencies using npm"""
-    try:
-        print('Installing Sage Web interface dependencies using npm...')
-        path = "{}/sage/http_server/static".format(getcwd())
-        run(["npm", "install", "--production"], cwd=path)
-        print('Sage Web interface successfully installed')
-    except SubprocessError as e:
-        print('Error: cannot install Sage Web interface successfully installed')
-        print('Error: {}'.format(e))
-        exit(1)
-
-
-class PostDevelopCommand(develop):
-    """Post-installation for development mode."""
-
-    def run(self):
-        install_web_deps()
-        develop.run(self)
-
-
-class PostInstallCommand(install):
-    """Post-installation for installation mode."""
-
-    def run(self):
-        install_web_deps()
-        print("moo")
-        install.run(self)
-
 
 __version__ = "2.0.1"
 
@@ -49,11 +15,6 @@ HDT_DEPS = [
 # dependencies required for the PostgreSQL backend
 POSTGRESQL_DEPS = [
     'psycopg2-binary==2.7.7'
-]
-
-# dependencies required for the CLI utilities (sage-query, sage-put, etc)
-COMMONS_DEPS = HDT_DEPS + POSTGRESQL_DEPS + [
-    'requests==2.21.0'
 ]
 
 console_scripts = [
@@ -84,15 +45,10 @@ setup(
     include_package_data=True,
     zip_safe=False,
     packages=find_packages(exclude=["tests", "tests.*"]),
-    cmdclass={
-        'develop': PostDevelopCommand,
-        'install': PostInstallCommand,
-    },
     # extras dependencies for the native backends
     extras_require={
         'hdt': HDT_DEPS,
-        'postgres': POSTGRESQL_DEPS,
-        'commons': COMMONS_DEPS
+        'postgres': POSTGRESQL_DEPS
     },
     entry_points={
         'console_scripts': console_scripts
